workers/requirements.txt
pandas
psycopg2
sqlalchemy
openpyxl
python-decouple==3.6
